Latest Java, peg the version of Jenkins to match videos
It seems the version of Java in the original will not download.  This updates the version of Java.  It also installs the specific version of Jenkins in the videos, otherwise there is a totally new way to get into the server that does not match the video.

The downside is it takes 30 minutes for the old version of jenkins to download - which meant the script timed out in my case, so I ran the jenkins install manually.  So the script updates would work if it would wait 30 minutes for jenkins.